+++ mopidy_subidy/backend.py
@@ -16,6 +16,7 @@ class SubidyBackend(pykka.ThreadingActor, backend.Backend):
             app_name=mopidy_subidy.SubidyExtension.dist_name,
             legacy_auth=subidy_config["legacy_auth"],
             api_version=subidy_config["api_version"],
+            radio_size=subidy_config["radio_size"],
         )
         self.library = library.SubidyLibraryProvider(backend=self)
         self.playback = playback.SubidyPlaybackProvider(